Littlefs working on flash!
Lots, lots of changes to make it happen. I hate doing this in one big commit, but it's sorta an all or nothing.

Fleshed out the settings to a few more things that will make sense. I don't love doing the raw structure read/write, but My other solution is an INI file and base64ing the binary bits, and that's just too much effort for now. Did find a decent library or two, we shall see. I'll probably still end up going with a MsgPack library of sorts eventually, unless there's a nice c json that does binary out of the box, so I don't have to do base64 roundtrips.

+++ src/files/files_fat32.h
@@ -26,7 +26,7 @@ namespace files
 		virtual bool Mount() override;
 		virtual bool Unmount() override;
 
-		virtual bool FileOpen(Fat32File& file, const char* filePath, uint8_t mode) override;
+		virtual bool FileOpen(Fat32File& file, const char* filePath, uint16_t mode) override;
 		virtual bool FileRead(Fat32File& file, void* buffer, size_t amountToRead, size_t* amountRead) override;
 		virtual bool FileWrite(Fat32File& file, const void* buffer, size_t bufferLength, size_t* amountWritten) override;
 		virtual bool FileSeek(Fat32File& file, size_t offset) override;
